requirements.txt
Mezzanine==3.0.6
psycopg2==2.5.2
Werkzeug==0.9.4
django-extensions==1.3.3
Pillow==2.3.0
Fabric==1.8.1
South==0.8.4
django-templated-email==0.4.9
pygeoip==0.3.0
